Refactor nav to remove react-toolbox, except search
